Change home design
