+++ app/src/main/AndroidManifest.xml
@@ -26,10 +26,6 @@
             android:exported="false"
             android:parentActivityName="com.sdp.swiftwallet.presentation.main.MainActivity" />
         <activity
-            android:name="com.sdp.swiftwallet.presentation.wallet.WalletInfoActivity"
-            android:exported="false"
-            android:parentActivityName="com.sdp.swiftwallet.presentation.main.MainActivity" />
-        <activity
             android:name="com.sdp.swiftwallet.presentation.cryptoGraph.CryptoGraphActivity"
             android:exported="false"
             android:parentActivityName="com.sdp.swiftwallet.presentation.cryptoGraph.CryptoValuesActivity" />
