+++ app/src/main/res/drawable/box.xml
@@ -5,11 +5,11 @@
 
     // The width and color of the border
     <stroke
-        android:width="3dp"
+        android:width="1dp"
         android:color="#c9c9c9" />
 
     // The corner radius, reduce it to make it more square
-    <corners android:radius="10dp"/>
+    <corners android:radius="2dp"/>
     <solid
         android:color="#f7f7f7"
         />
