+++ app/src/main/res/drawable/round.xml
@@ -6,16 +6,19 @@
     // The width and color of the border
     <stroke
         android:width="2dp"
-        android:color="#428dff" />
+        android:color="@color/primary" />
 
     // The corner radius, reduce it to make it more square
-    <corners android:radius="1dp"/>
+    <corners android:radius="300dp"/>
     <solid
-        android:color="@color/white"
+        android:color="#f7f7f7"
         />
+    // Add your desired padding
     <padding
-        android:left="10dp"
-        android:top="5dp"
-        android:right="10dp"
-        android:bottom="5dp" />
+        android:left="20dp"
+        android:top="10dp"
+        android:right="20dp"
+        android:bottom="10dp" >
+
+    </padding>
 </shape>
