--- app/src/main/res/layout/activity_wallet_info.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context="com.sdp.swiftwallet.presentation.wallet.WalletInfoActivity">
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        android:orientation="vertical"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        android:id="@+id/wallet_info_layout"
-        app:layout_constraintBottom_toTopOf="@id/qr_info_view"
-        android:layout_marginTop="@dimen/activity_vertical_margin"
-        >
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:text="@string/address_label_text"/>
-        <TextView
-            android:id="@+id/wallet_info_address"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:textColor="@color/black"
-            android:textSize="@dimen/wallet_text_highlight"
-            android:text="@string/random_example_address"
-            android:textIsSelectable="true"/>
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:text="@string/balance_label_text"/>
-        <TextView
-            android:id="@+id/wallet_info_balance"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAlignment="center"
-            android:textColor="@color/black"
-            android:textSize="@dimen/wallet_text_highlight"
-            android:text="@string/random_example_balance"/>
-    </LinearLayout>
-    <ImageView
-        android:id="@+id/qr_info_view"
-        android:layout_width="268dp"
-        android:layout_height="301dp"
-        android:layout_marginTop="@dimen/activity_vertical_margin"
-        app:layout_constraintTop_toBottomOf="@+id/wallet_info_layout"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:srcCompat="@drawable/ether"
-        />
-</androidx.constraintlayout.widget.ConstraintLayout>
